+++ src/commands/toggleAlphabeticalSounds.ts
@@ -2,15 +2,15 @@ import * as vscode from 'vscode';
 import { settings } from '../settings/pluginSettings';
 import { setConfig } from '../settings/configState';
 
-export function getToggleSFXCommand() {
-    const toggleSFXCmd = vscode.commands.registerCommand('vscode-animalese.toggleSFX', () => {
-        const newValue = !settings.onlySFX;
-        setConfig('onlySFX', newValue);
+export function getToggleSoundsCommand() {
+    const toggleSoundsCmd = vscode.commands.registerCommand('vscode-animalese.toggleSounds', () => {
+        const newValue = !settings.alphabeticalSounds;
+        setConfig('alphabeticalSounds', newValue);
         vscode.window.showInformationMessage(
             newValue
                 ? 'Disabled individual sounds for alphabetical keys.'
                 : 'Enabled individual sounds for alphabetical keys.',
         );
     });
-    return toggleSFXCmd;
+    return toggleSoundsCmd;
 }