+++ src/commands/toggleExtension.ts
@@ -1,0 +1,11 @@
+import * as vscode from 'vscode';
+import { extensionEnabled, setExtensionEnabled } from '../extension';
+import { disablingText, enablingText } from '../constants/popupText';
+
+export function getToggleExtensionCommand() {
+    const toggleCmd = vscode.commands.registerCommand('vscode-animalese.toggle', () => {
+        setExtensionEnabled(!extensionEnabled);
+        vscode.window.showInformationMessage(extensionEnabled ? enablingText : disablingText);
+    });
+    return toggleCmd;
+}