+++ src/get/cacheString.ts
@@ -4,5 +4,5 @@ import { settings } from '../settings/pluginSettings';
 export default function getCacheString(key: string) {
     return `${key}${VOICE_LIST.indexOf(
         settings.voice,
-    )}${+settings.specialPunctuation}${+settings.onlySFX}${+settings.harmonicSFX}${+settings.diacriticRecognition}`;
+    )}${+settings.specialPunctuation}${+settings.alphabeticalSounds}${+settings.harmonicSounds}${+settings.diacriticRecognition}`;
 }