+++ CHANGELOG.md
@@ -2,8 +2,8 @@
 
 ## v0.2.3
 
-- Added `onlySFX` configuration to only play basic sound effects without vocal sounds for alphabetical characters.
-- Added `harmonicSFX` configuration to toggle harmonic sounds for numbers and `-`/`=` characters. This addition and the above addition are courtesy of [KaeruDev](https://github.com/KaeruDev).
+- Added the `alphabeticalSounds` configuration to toggle individual sounds for alphabetical characters.
+- Added the `harmonicSounds` configuration to toggle harmonic sounds for numbers and `-`/`=` characters. This addition and the above addition are courtesy of [KaeruDev](https://github.com/KaeruDev).
 - Added `diacriticRecognition` configuration to toggle sound effects for characters with various diacritics (thanks to [RobotoSkunk](https://github.com/RobotoSkunk)) for the suggestion!
 
 ## v0.2.2